Patch for the 2 new bosses Nolla added
This adds the two bosses that Nolla added into Purgatory. I haven't done their reworks (obviously), but their vanilla forms are now in the mod. This way you can kill the meat boss and deactivate the curse. Sauvan Ydin is also a work in progress during this patch, so feel free to mess with its experimental gui stage.